fix: cfloat include edits
